Problem with persistence
There seems to be a problem with the persistence. The Spatial View
doesn't return the full 10000 documents when requested first, it
only returns the correct result, when requestes a second time.

It should work on the first request as there is an explicit
self.wait_for_persistence() call in
SpatialHelper.query_index_for_verification()

I see this behaviour on my machine when I run 4 nodes with
./cluster_run -n 4 and the test
test_spatial_view_on_x_docs_y_design_docs (others are commented out)
with
./testrunner -i b/resources/dev-4-nodes.ini -c conf/py-spatial.conf

It does not happen if I reduce the number of inserted documents or
the number design documents. Hence I think the data gets inserted
but the disk is not fast enough to persist them, although the
stats say they are persited.

Change-Id: I867e53698f0d4c1015210dcdca85b11494f2305c